tests/wpt/web-platform-tests/css/css-anchor-position/anchor-position-multicol-002.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-pos">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.cb {
  transform: translate(0);  /* Make it a containing block. */
  border-width: 5px 6px 7px 8px;
  border-style: solid;
  padding: 5px 6px 7px 8px;
}
.columns {
  column-count: 4;
  column-fill: auto;
  column-gap: 10px;
  column-width: 100px;
  width: 320px;
  height: 100px;
}
.spacer {
  height: 10px;
  background: pink;
}
.anchor1 {
  anchor-name: --a1;
  margin-left: 10px;
  width: 50px;
  height: 90px;
  background: blue;
}
.target {
  position: absolute;
}
.fixed {
  position: fixed;
}
.target1 {
  left: anchor(--a1 left);
  top: anchor(--a1 top);
  width: anchor-size(--a1 width);
  height: anchor-size(--a1 height);
  background: lime;
  opacity: .3;
}
.target1-rb {
  right: anchor(--a1 right);
  bottom: anchor(--a1 bottom);
  width: 10px;
  height: 10px;
  background: purple;
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <p>The green box should be a union of blue boxes,
    and the purple box should be at the right-bottom of the green box.
  </p>
  <div class="spacer" style="height: 10px"></div>
  <div class="cb">
    <div class="columns">
      <div class="spacer" style="height: 30px"></div>
      <div class="cb">
        <!-- This spacer fills up to the middle of the 2nd column. -->
        <div class="spacer" style="height: 130px"></div>
        <div class="anchor1"></div>
        <div class="spacer" style="height: 100px"></div>

        <!-- The containing block of querying elements is block-fragmented. -->
        <div class="target target1"
             data-offset-x=18 data-offset-y=65
             data-expected-width=160 data-expected-height=100></div>
        <div class="target target1-rb"
             data-offset-x=168 data-offset-y=155></div>
        <div class="target fixed target1"
             data-offset-x=26 data-offset-y=70
             data-expected-width=160 data-expected-height=100></div>
        <div class="target fixed target1-rb"
             data-offset-x=176 data-offset-y=160></div>
      </div>

      <!-- The containing block of querying elements is a multi-column.  -->
      <div class="target target1"
           data-offset-x=144 data-offset-y=5
           data-expected-width=160 data-expected-height=100></div>
      <div class="target target1-rb"
           data-offset-x=294 data-offset-y=95></div>
      <div class="target fixed target1"
           data-offset-x=152 data-offset-y=10
           data-expected-width=160 data-expected-height=100></div>
      <div class="target fixed target1-rb"
           data-offset-x=302 data-offset-y=100></div>
    </div>

    <!-- The containing block of querying elements is not fragmented.  -->
    <div class="target target1"
         data-offset-x=144 data-offset-y=5
         data-expected-width=160 data-expected-height=100></div>
    <div class="target target1-rb"
         data-offset-x=294 data-offset-y=95></div>
    <div class="target fixed target1"
         data-offset-x=152 data-offset-y=10
         data-expected-width=160 data-expected-height=100></div>
    <div class="target fixed target1-rb"
         data-offset-x=302 data-offset-y=100></div>
    </div>
</body>
